Add couple more days, hide in utils
