+++ multiqc/utils/config_defaults.yaml
@@ -22,7 +22,6 @@ custom_logo: null
 custom_logo_url: null
 custom_logo_title: null
 custom_css_files: []
-easter_eggs: true
 simple_output: false
 template: "default"
 profile_runtime: false
